Make dcape_dcin a method of Environment
