+++ dparcel/thermo.py
@@ -15,8 +15,6 @@ import metpy.constants as const
 from metpy.units import units, concatenate
 
 from scipy.special import lambertw
-from scipy.optimize import minimize_scalar
-from scipy.integrate import simps
 
 
 # ---------- Basic thermodynamic calculations ----------
@@ -240,64 +238,6 @@ def saturation_equivalent_potential_temperature(p, Tk, prime=False):
     )
 
     return thetae, thetae*dlogthetae_dTk
-
-
-def dcape_dcin(sounding, samples=10000):
-    """
-    Compute DCAPE and DCIN for a sounding according to Market et. al. (2017).
-
-    Args:
-        sounding: An Environment object.
-        samples: Number of samples to use for integration (optional).
-
-    Returns:
-        DCAPE and DCIN for the sounding.
-
-    References:
-        Market, PS, Rochette, SM, Shewchuk, J, Difani, R, Kastman, JS,
-        Henson, CB & Fox, NI 2017, ‘Evaluating elevated convection
-        with the downdraft convective inhibition’, Atmospheric
-        science letters, vol. 18, no. 2, pp. 76–81.
-    """
-    # find minimum wet bulb temperature in lowest 6 km
-    def env_wetbulb(z):
-        return sounding.wetbulb_temperature(z*units.meter).m_as(units.kelvin)
-    sol = minimize_scalar(env_wetbulb, bounds=(0, 6000), method='bounded')
-    z_initial = sol.x
-    p_initial = sounding.pressure(z_initial*units.meter)
-    t_initial = sol.fun*units.kelvin
-
-    def integrand(z_final):
-        # find the virtual temperature after moist pseudoadiabatic
-        # descent to the final level
-        z_final = z_final*units.meter
-        p_final = sounding.pressure(z_final)
-        t_final = moist_lapse(
-            p_final, t_initial, p_initial, method='integration')
-        w_final = mpcalc.saturation_mixing_ratio(p_final, t_final)
-        tv_final = mpcalc.virtual_temperature(t_final, w_final)
-
-        # find the environmental virtual temperature at that level
-        t_env = sounding.temperature(z_final)
-        w_env = mpcalc.mixing_ratio_from_specific_humidity(
-            sounding.specific_humidity(z_final))
-        tv_env = mpcalc.virtual_temperature(t_env, w_env)
-
-        return 1 - tv_final.m_as(units.kelvin)/tv_env.m_as(units.kelvin)
-
-    # DCAPE: integrate from surface to level of minimum wet bulb
-    # temperature, taking positive area only.
-    # passing the integrand function to scipy.integrate.quad is very
-    # slow so we compute many samples and use Simpson's method.
-    z_sample = np.linspace(0, z_initial, samples)
-    dcape = simps(
-        np.maximum(integrand(z_sample), 0), z_sample)*units.meter*const.g
-    # DCIN: integrate from surface to level of minimum wet bulb
-    # temperature, taking negative area only
-    dcin = simps(
-        np.minimum(integrand(z_sample), 0), z_sample)*units.meter*const.g
-
-    return dcape, dcin
 
 
 def lcl_romps(p, T, q):
